refactor: update graph view presentation component and import structure
