+++ components/graph-view/container/graph-view.tsx
@@ -2,7 +2,7 @@ import { getPopulation } from "@/api/getPopulation";
 import { getPrefectures } from "@/api/getPrefectures";
 import { searchParamsCache } from "@/lib/search-params";
 import type { FC } from "react";
-import { ClientOnlyGraphViewPresentation as GraphViewPresentation } from "../presentation/wrapper";
+import { GraphViewPresentation } from "../presentation/graph-view-presentation";
 
 export const GraphView: FC = async () => {
 	const searchParams = searchParamsCache.all();
